test: refactor helpers & setup into external file
